requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
crispy-bootstrap5==2024.2
Django==4.2.13
django-bootstrap-v5==1.0.11
django-crispy-forms==2.1
python-dateutil==2.9.0.post0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
